mumblemumble
doesn't hurt

+++ src/wp-includes/class-phpass.php
@@ -173,7 +173,7 @@ class PasswordHash {
 		$itoa64 = './ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
 
 		$output = '$2a$';
-		$output .= chr(ord('0') + $this->iteration_count_log2 / 10);
+		$output .= chr((int)(ord('0') + $this->iteration_count_log2 / 10));
 		$output .= chr((int)(ord('0') + $this->iteration_count_log2 % 10));
 		$output .= '$';
